refactor: :recycle: 아바타, 닉네임 변경 컴포넌트 분리
재활용 가능해보여서
