+++ frontend/src/common/ChangeProfiles/index.ts
@@ -1,0 +1,2 @@
+export { ChangeAvatar } from './ChangeAvatar';
+export { ChangeNickname } from './ChangeNickname';
