+++ frontend/src/common/index.ts
@@ -4,3 +4,4 @@ export { Modal } from './Modal';
 export { Dropdown } from './Dropdown';
 export { Tooltip } from './Tooltip';
 export { RootComponent } from './RootComponent';
+export { ChangeAvatar, ChangeNickname } from './ChangeProfiles';
